+++ src/app/analyze/keyword/page.tsx
@@ -8,6 +8,7 @@ import useInput from '@/hooks/useInput';
 import TextAreaField from '@/components/Input/TextAreaField/TextAreaField';
 import { exceptEnter } from '@/shared/utils/exceptEnter';
 import Button from '@/components/Button/Button';
+import type { Keyword } from '@/feature/analyze/types';
 
 // FIXME: mock data
 const data = {
@@ -37,7 +38,7 @@ const data = {
 };
 const entriesData = Object.entries(data);
 
-const deDuplicatedKeywordList = (arr: [string, boolean][]) => Object.entries(Object.fromEntries(arr));
+const deDuplicatedKeywordList = (arr: Keyword) => Object.entries(Object.fromEntries(arr));
 
 // TODO: 경험 분해 키워드 가져오기 API (/experience/capability/{experienceId}) 요청으로 키워드 불러오기
 // TODO: 경험 분해 키워드 임시 저장 API (/experience/capability) 로 요청 보내기, 요청body: Object.fromEntries(keywordList.filter((keyword) => keyword[1] === true))
@@ -54,7 +55,7 @@ const Keyword = () => {
       const toggleKeyword = selectedKeywordList.includes(keywordList[selectedKeywordIndex]);
       const newKeywordList = prev.map(([prevKeyword, prevSelected]) =>
         prevKeyword === newKeyword ? [prevKeyword, !prevSelected] : [prevKeyword, prevSelected]
-      ) as [string, boolean][];
+      ) as Keyword;
 
       if (selectedKeywordList.length === 4 && toggleKeyword) return newKeywordList;
       if (selectedKeywordList.length >= 4) return prev;
